+++ test/node/sanitizations.spec.js
@@ -206,3 +206,23 @@ test.group('Sanitizations | toNull', () => {
     assert.isArray(sanitizations.toNull([]))
   })
 })
+
+test.group('Sanitizations | escape', () => {
+  test('escape HTML string', (assert) => {
+    assert.equal(sanitizations.escape('<div> hello </div>'), '&lt;div&gt; hello &lt;/div&gt;')
+  })
+
+  test('return same value when it\'s not a string', (assert) => {
+    assert.equal(sanitizations.escape(22), 22)
+  })
+})
+
+test.group('Sanitizations | stripLinks', () => {
+  test('remove anchor tags from a string', (assert) => {
+    assert.equal(sanitizations.stripLinks('<a href=""> google.com </a>'), 'google.com')
+  })
+
+  test('return same value when it\'s not a string', (assert) => {
+    assert.equal(sanitizations.stripLinks(22), 22)
+  })
+})